requirements.txt
Django>=3.0.4
djangorestframework>=3.11.0
pandas>1.0.0